+++ cookowly-frontend/.eslintrc.json
@@ -45,6 +45,7 @@
     ],
     "function-paren-newline": "off",
     "no-shadow": "off",
-    "@typescript-eslint/no-shadow": "error"
+    "@typescript-eslint/no-shadow": "error",
+    "react-hooks/exhaustive-deps": "warn"
   }
 }
